Fix size_t-int conversion warning.

+++ src/Core/TundraCore/Framework/Framework.cpp
@@ -627,7 +627,7 @@ bool Framework::LoadStartupOptionsFromXML(QString configurationFile)
 
 void Framework::AddCommandLineParameter(const QString &command, const QString &parameter)
 {
-    startupOptions.insert(std::make_pair(command, std::make_pair(startupOptions.size() + 1, parameter)));
+    startupOptions.insert(std::make_pair(command, std::make_pair((int)startupOptions.size() + 1, parameter)));
 }
 
 bool Framework::HasCommandLineParameter(const QString &value) const
